Update README.md
A regression algorithm is a type of supervised machine learning algorithm that models the relationship between a dependent variable (target) and one or more independent variables (features). The goal of regression is to find the best-fit line or curve that represents the relationship between the variables, allowing us to make predictions or understand the underlying patterns in the data.